+++ djangocms_translations/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '1.3.4'
+__version__ = '1.3.6'
 
 default_app_config = 'djangocms_translations.apps.DjangocmsTranslationsConfig'