+++ djangocms_translations/providers/supertext.py
@@ -17,6 +17,30 @@ from ..utils import add_domain, get_translatable_fields, get_text_field_child_la
 from .base import BaseTranslationProvider, ProviderException
 
 
+# Format: language_code used in settings.LANGUAGE_* --> language_code that will be sent to Supertext
+LANGUAGE_MAPPING = {
+    'ch-de': 'de-CH',
+    'ch-fr': 'fr-CH',
+    'da': 'da-DK',
+    'de': 'de-DE',
+    'en': 'en-US',
+    # 'en-nz': FIXME: Supertext does not support this language
+    # 'en-za': FIXME: Supertext does not support this language
+    'es': 'es-ES',
+    'es-xl': 'es-419',
+    'fi': 'fi-FI',
+    'fr': 'fr-FR',
+    'it': 'it-IT',
+    'ja': 'ja-JP',
+    'nb': 'nb-NO',
+    'nl': 'nl-NL',
+    'pl': 'pl-PL',
+    'ru': 'ru-RU',
+    'sk': 'sk-SK',
+    'sv': 'sv-SE',
+}
+
+
 def _get_translation_export_content(field, raw_plugin):
     plugin_class = get_plugin_class(raw_plugin['plugin_type'])
     try:
@@ -85,8 +109,8 @@ class SupertextTranslationProvider(BaseTranslationProvider):
     def get_export_data(self):
         data = {
             'ContentType': 'text/html',
-            'SourceLang': self.request.source_language,
-            'TargetLanguages': [self.request.target_language],
+            'SourceLang': LANGUAGE_MAPPING.get(self.request.source_language, self.request.source_language),
+            'TargetLanguages': [LANGUAGE_MAPPING.get(self.request.target_language, self.request.target_language)],
         }
         groups = []
         fields_by_plugin = {}